requirements.txt
# ── IMS AstroBot Dependencies ──

# Streamlit UI
streamlit>=1.30.0

# LLM via llama.cpp
llama-cpp-python>=0.2.50

# Embeddings
sentence-transformers>=2.2.0

# Vector database
chromadb>=0.4.22

# Document parsers
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Text processing
tiktoken>=0.5.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.1.0
